+++ cmd/clay/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"embed"
+
 	"github.com/go-go-golems/clay/cmd/clay/repo"
 	clay "github.com/go-go-golems/clay/pkg"
 	"github.com/go-go-golems/glazed/pkg/cli"